Setup Flyway Migration, CustomerJDBCDataAccessServiceTest.

+++ backend/src/main/java/com/anagracetech/backend/customer/Customer.java
@@ -33,7 +33,7 @@ public class Customer implements UserDetails {
     @GeneratedValue(
             strategy = GenerationType.SEQUENCE,
             generator = "customer_id_seq")
-    private Long id;
+    private Integer id;
 
     @Column(
             nullable = false
@@ -67,7 +67,7 @@ public class Customer implements UserDetails {
     public Customer() {
     }
 
-    public Customer(Long id,
+    public Customer(Integer id,
                     String name,
                     String email,
                     String password,
@@ -81,7 +81,7 @@ public class Customer implements UserDetails {
         this.gender = gender;
     }
 
-    public Customer(Long id,
+    public Customer(Integer id,
                     String name,
                     String email,
                     String password,
@@ -104,11 +104,11 @@ public class Customer implements UserDetails {
         this.gender = gender;
     }
 
-    public Long getId() {
+    public Integer getId() {
         return id;
     }
 
-    public void setId(Long id) {
+    public void setId(Integer id) {
         this.id = id;
     }
